refactor: Split `/hooks/` into domains
This will make adding mutations cleaner later
