+++ src/app/tournament/[uuid]/_components/DeletePlayerModal.tsx
@@ -4,7 +4,7 @@ import { FilledButton } from "@/elements/FilledButton";
 import { Modal } from "@/elements/Modal";
 import { Notification } from "@/elements/Notification";
 import { RoundedListbox } from "@/elements/RoundedListbox";
-import { usePlayers } from "@/hooks/usePlayers";
+import { usePlayers } from "@/hooks/players/usePlayers";
 import type { Player } from "@/types/app.types";
 
 type DeletePlayerModalProps = {
