+++ src/app/tournament/[uuid]/_components/EditPlayerModal.tsx
@@ -5,7 +5,7 @@ import { LabelledInput } from "@/elements/LabelledInput";
 import { Modal } from "@/elements/Modal";
 import { Notification } from "@/elements/Notification";
 import { RoundedListbox } from "@/elements/RoundedListbox";
-import { usePlayers } from "@/hooks/usePlayers";
+import { usePlayers } from "@/hooks/players/usePlayers";
 import { parseFormString } from "@/lib/utils";
 import type { Player } from "@/types/app.types";
