+++ src/app/tournament/[uuid]/_components/PlayerList.tsx
@@ -6,9 +6,9 @@ import {
   unlockPlayer,
 } from "@/actions/attendance";
 import { IconButton } from "@/elements/IconButton";
-import { useAttendance } from "@/hooks/useAttendance";
-import { useLogs } from "@/hooks/useLogs";
-import { useTables } from "@/hooks/useTables";
+import { useAttendance } from "@/hooks/attendance/useAttendance";
+import { useLogs } from "@/hooks/logs/useLogs";
+import { useTables } from "@/hooks/tables/useTables";
 import { rankPlayers, scoreToColor } from "@/lib/scores";
 import type { Player, Session } from "@/types/app.types";
