+++ src/app/tournament/[uuid]/_components/Sidebar.tsx
@@ -11,9 +11,9 @@ import { registerPlayer } from "@/actions/attendance";
 import { FilledButton } from "@/elements/FilledButton";
 import { IconButton } from "@/elements/IconButton";
 import { SearchCombobox } from "@/elements/SearchCombobox";
-import { useAttendance } from "@/hooks/useAttendance";
-import { usePlayers } from "@/hooks/usePlayers";
-import { useSessions } from "@/hooks/useSessions";
+import { useAttendance } from "@/hooks/attendance/useAttendance";
+import { usePlayers } from "@/hooks/players/usePlayers";
+import { useSessions } from "@/hooks/sessions/useSessions";
 import type { Player } from "@/types/app.types";
 import { CreatePlayerModal } from "./CreatePlayerModal";
 import { DeletePlayerModal } from "./DeletePlayerModal";
