+++ src/app/tournament/[uuid]/_components/TableSeat.tsx
@@ -2,9 +2,9 @@ import { useEffect, useState } from "react";
 import { twMerge } from "tailwind-merge";
 import { updateTable } from "@/actions/tables";
 import { RoundedListbox } from "@/elements/RoundedListbox";
-import { useAttendance } from "@/hooks/useAttendance";
-import { usePlayers } from "@/hooks/usePlayers";
-import { useTables } from "@/hooks/useTables";
+import { useAttendance } from "@/hooks/attendance/useAttendance";
+import { usePlayers } from "@/hooks/players/usePlayers";
+import { useTables } from "@/hooks/tables/useTables";
 import { useTournaments } from "@/hooks/useTournaments";
 import { getPointDeltas } from "@/lib/scores";
 import { useTournamentContext } from "@/providers/TournamentProvider";
