+++ src/app/tournament/[uuid]/analytics/_components/PlayerAnalytics.tsx
@@ -1,5 +1,5 @@
 import { LineGraph } from "@/elements/LineGraph";
-import { useLogs } from "@/hooks/useLogs";
+import { useLogs } from "@/hooks/logs/useLogs";
 import { useTournaments } from "@/hooks/useTournaments";
 import { getPointHistory } from "@/lib/scores";
 import { useTournamentContext } from "@/providers/TournamentProvider";
