+++ src/app/tournament/[uuid]/logs/_components/LogList.tsx
@@ -2,8 +2,8 @@ import { Trash2 } from "lucide-react";
 import { twMerge } from "tailwind-merge";
 import { disableLog } from "@/actions/logs";
 import { IconButton } from "@/elements/IconButton";
-import { usePlayers } from "@/hooks/usePlayers";
-import { useSessions } from "@/hooks/useSessions";
+import { usePlayers } from "@/hooks/players/usePlayers";
+import { useSessions } from "@/hooks/sessions/useSessions";
 import { WIN_TYPE_MAP } from "@/lib/constants";
 import type { Log } from "@/types/app.types";
