+++ src/app/tournament/[uuid]/page.tsx
@@ -5,9 +5,9 @@ import { useState, useTransition } from "react";
 import { shuffleTables } from "@/actions/tables";
 import { FilledButton } from "@/elements/FilledButton";
 import { RoundedListbox } from "@/elements/RoundedListbox";
-import { useAttendance } from "@/hooks/useAttendance";
-import { useSessions } from "@/hooks/useSessions";
-import { useTables } from "@/hooks/useTables";
+import { useAttendance } from "@/hooks/attendance/useAttendance";
+import { useSessions } from "@/hooks/sessions/useSessions";
+import { useTables } from "@/hooks/tables/useTables";
 import { WIND_MAP, WINDS } from "@/lib/constants";
 import { Sidebar } from "./_components/Sidebar";
 import { TableList } from "./_components/TableList";
