+++ src/app/tournament/[uuid]/sessions/page.tsx
@@ -4,9 +4,9 @@ import { ChartColumn } from "lucide-react";
 import { useState } from "react";
 import { FilledButton } from "@/elements/FilledButton";
 import { RoundedListbox } from "@/elements/RoundedListbox";
-import { useLogs } from "@/hooks/useLogs";
-import { usePlayers } from "@/hooks/usePlayers";
-import { useSessions } from "@/hooks/useSessions";
+import { useLogs } from "@/hooks/logs/useLogs";
+import { usePlayers } from "@/hooks/players/usePlayers";
+import { useSessions } from "@/hooks/sessions/useSessions";
 import type { Session } from "@/types/app.types";
 import { Leaderboard } from "./_components/Leaderboard";
 import { ViewGraphModal } from "./_components/ViewGraphModal";
